- debugged a few regression datasets (their targets were converted to ints) - debugged mlpython_helper (was importing the wrong mlpython) - debugged unit tests (a few were failing, for different reasons)

+++ datasets/cadata.py
@@ -61,7 +61,7 @@ def load(dir_path,load_to_memory=False):
 
     lengths = [16512, 2064, 2064]
     if load_to_memory:
-        train,valid,test = [mlio.MemoryDataset(d,[(input_size,),(1,)],[np.float64,int],l) for d,l in zip([train,valid,test],lengths)]
+        train,valid,test = [mlio.MemoryDataset(d,[(input_size,),(1,)],[np.float64,np.float64],l) for d,l in zip([train,valid,test],lengths)]
         
     # Get metadata
     train_meta,valid_meta,test_meta = [{'input_size':input_size, 'length':l} for l in lengths]
